+++ src/helpers.php
@@ -2,7 +2,7 @@
 
 namespace Kalnoy\Cruddy;
 
-if (!function_exists('Kalnoy\Cruddy\prettify_string'))
+if ( ! function_exists('Kalnoy\Cruddy\prettify_string'))
 {
     /**
      * @param $value
@@ -15,7 +15,7 @@ if (!function_exists('Kalnoy\Cruddy\prettify_string'))
     }
 }
 
-if (!function_exists('Kalnoy\Cruddy\try_trans'))
+if ( ! function_exists('Kalnoy\Cruddy\try_trans'))
 {
     /**
      * Try translate a key.
@@ -26,11 +26,11 @@ if (!function_exists('Kalnoy\Cruddy\try_trans'))
      */
     function try_trans($key)
     {
-        return strpos($key, '.') !== false ? trans($key) : $key;
+        return app('cruddy.lang')->tryTranslate($key);
     }
 }
 
-if (!function_exists('Kalnoy\Cruddy\ucfirst'))
+if ( ! function_exists('Kalnoy\Cruddy\ucfirst'))
 {
     /**
      * `uncfirst` for unicode strings.