+++ CHANGELOG.md
@@ -6,3 +6,6 @@ __0.3.0__
 *   Removed `DateTime::format()`, the format is fixed now
 *   A better date and time inputs
 *   Fix issues with entity selector
+*   Move lang-related stuff to `cruddy.lang` service
+*   Move assets-related stuff to `cruddy.assets` service
+*   `SchemaInterface::entity` now doesn't have any parameters